feat: centralize cart API handler
