+++ apps/shop-abc/__tests__/cartApi.test.ts
@@ -5,13 +5,11 @@ import {
 } from "@platform-core/src/cartCookie";
 import { createCart, getCart, setCart } from "@platform-core/src/cartStore";
 import { PRODUCTS } from "@platform-core/products";
-import crypto from "crypto";
 import { DELETE, GET, PATCH, POST } from "../src/app/api/cart/route";
 import * as productLib from "@/lib/products";
 
 const TEST_SKU = PRODUCTS[0];
 
-process.env.CART_COOKIE_LEGACY_SECRET ||= "legacy-secret";
 
 // Minimal NextResponse mock using the native Response class
 jest.mock("next/server", () => ({
@@ -160,36 +158,4 @@ test("GET returns cart", async () => {
   const res = await GET(createRequest({}, encodeCartCookie(cartId)));
   const body = await res.json();
   expect(body.cart).toEqual(cart);
-});
-
-test("PATCH rejects unsigned legacy cookie", async () => {
-  const sku = { ...TEST_SKU, id: "01ARZ3NDEKTSV4RRFFQ69G5FAA" };
-  const size = sku.sizes[0];
-  const lineId = `${sku.id}:${size}`;
-  const cart = { [lineId]: { sku, qty: 1, size } };
-  const legacyPayload = Buffer.from(JSON.stringify(cart)).toString("base64");
-  // Missing signature
-  const res = await PATCH(createRequest({ id: lineId, qty: 2 }, legacyPayload));
-  expect(res.status).toBe(404);
-});
-
-test("GET migrates signed legacy cookie", async () => {
-  const sku = { ...TEST_SKU, id: "01ARZ3NDEKTSV4RRFFQ69G5FAA" };
-  const size = sku.sizes[0];
-  const lineId = `${sku.id}:${size}`;
-  const cart = { [lineId]: { sku, qty: 2, size } };
-  const payload = Buffer.from(JSON.stringify(cart)).toString("base64");
-  const sig = crypto
-    .createHmac("sha256", process.env.CART_COOKIE_LEGACY_SECRET!)
-    .update(payload)
-    .digest("hex");
-  const legacyCookie = `${payload}.${sig}`;
-  const res = await GET(createRequest({}, legacyCookie));
-  const body = await res.json();
-  expect(body.cart).toEqual(cart);
-  const header = res.headers.get("Set-Cookie")!;
-  const encoded = header.split(";")[0].split("=")[1];
-  const id = decodeCartCookie(encoded)!;
-  const stored = await getCart(id);
-  expect(stored).toEqual(cart);
 });
